device: add a test that demonstrates handshakes racing
Each side attempts to establish a tunnel at the same time.
The result is the handshakes cross in-flight, both sides get
confused, and we have to wait for the multi-second timeout to
get a tunnel established.

Without a handshake race, it is possible to set

	const maxWait = 300 * time.Millisecond

to get the test passing reliably even on slow machines.

Signed-off-by: David Crawshaw <crawshaw@tailscale.io>