[Experimental] Avoid handshake race when two peers initiate at once.
Node A would send HandshakeInitiation#1 to B, and B would send
HandshakeInitiation#2 to A, simultaneously. Then A would reply with
HandshakeResponse#2 and B would reply with HandshakeResponse#1, and so
on. The result is that the two nodes could each believe the handshake
had completed successfully, yet be in disagreement about the session
key.

To fix it, we arbitrarily choose that if device X has recently sent a
HandshakeInitiation, *and* its public key is !LessThan the peer's public
key, it will skip "consuming" an incoming initiation packet (other than to
update its endpoint information). It won't generate a HandshakeResponse
in this case. This allows X's original handshake process to complete.

In theory, this change should not break compatibility with older
wireguard implementations that don't implement this workaround (except
the race condition will only be 50% fixed when communicating with such
peers).

TODO:
- Should we immediately retransmit our own initiation packet as soon as
  getting an incoming initiation?
- Should we unconditionally discard incoming handshakes and always
  initiate our own? This would give fewer edge cases to worry about.
- Is there a chance our initiation will fail or give up for some
  reason, while the peer is still retrying?