[HACK] peer: Don't allow UpdateDst() to an endpoint in our list of routes.
Otherwise we could end up with a routing loop: trying to reach one of
our peers through itself.

TODO(apenwarr): Long-term this is not the right approach.
What we want is for wireguard to never send packets through itself,
even if the addresses overlap. For example, an endpoint could be
10.0.0.10, and that endpoint could provide routes to 10.0.0.0/24. When
wireguard tries to send to 10.0.0.10, it should go directly. When any
other process does, it should go through the tunnel. Original wireguard
is capable of this, but Tailscale has broken it at the moment. Once
we've fixed it (on both Windows and Linux) we can remove this hack.