device: don't call peer.Stop in unsafeRemovePeer
peer.Stop blocks on the peer.routines.stopping WaitGroup, which
waits for a goroutine which wants to call peer.timersActive,
taking a read lock on the device.peers mutex. The callers of
unsafeRemovePeer lock device.peers, so this deadlocks.

Two example stacks from the deadlock, the first is waiting
while holding the peers mutex:

	sync.runtime_Semacquire                        src/runtime/sema.go:56
	sync.(*WaitGroup).Wait                         src/sync/waitgroup.go:130
	wireguard-go/device.(*Peer).Stop               device/peer.go:312
	wireguard-go/device.unsafeRemovePeer           device/device.go:130
	wireguard-go/device.(*Device).RemoveAllPeers   device/device.go:443
	wireguard-go/device.(*Device).IpcSetOperation  device/uapi.go:212

The second wants the mutex:

	sync.(*RWMutex).RLock                                             src/sync/rwmutex.go:50
	wireguard-go/device.(*Peer).timersActive                          device/timers.go:78
	wireguard-go/device.(*Peer).timersAnyAuthenticatedPacketTraversal device/timers.go:223
	wireguard-go/device.(*Peer).SendHandshakeInitiation               device/send.go:166
	wireguard-go/device.expiredRetransmitHandshake                    device/timers.go:120
	wireguard-go/device.(*Peer).NewTimer.func1                        device/timers.go:42

Instead, call peer.Stop after the device.peers lock has been
released.

(This commit is a rewrite of work by Avery Pennarun.)

Signed-off-by: David Crawshaw <crawshaw@tailscale.com>