device: prevent more races on endpoint
The existing code attempted to prevent races on peer.endpoint
due to logging by adding a no-op String method to conn.NativeEndpoint.
However, the conn.NativeEndpoint got shoved inside a conn.Endpoint
interface, which meant that there was already a race trying to get
the type out in order to discover that printing it was a no-op.

There have been multiple attempts to fix this, in different ways.
We've hit it with both endpoints having a String methods and
an Endpoint type not having one (and thus fmt reflecting over all
fields of the Endpoint to print them, including a Endpoint's internal
mutex fields.)

Improve this bandaid with a big hammer: read peer.endpoint entirely
in the critical section.

If you find yourself touching this code, please run

go test -c -race && stress ./device.test -test.timeout=30s

for a while to make sure you do not accidentally re-introduce a race.

Even better yet would be to make conn.NativeEndpoint a struct
that does some internal synchronization, but that would be
a much more invasive change.