device: make CreateEndpoint have only one param
When opts are set, concatenate the public key and the addrs.
Making the function signature of DeviceOptions.createEndpoint match
that of conn.CreateEndpoint makes merges with upstream easier.
This is ugly, but it is also temporary.
I plan to embed the public key in the addr in a more principled
way in the future on the Tailscale side.
This just unblocks me while I work on other aspects
of merging and fork removal.

While we're here, make skipBindUpdate implicit as well.

Signed-off-by: Josh Bleecher Snyder <josharian@gmail.com>